relocacion de archivos, modularizacion panel derecho
