+++ src/components/Botones/BotonNavigate.jsx
@@ -1,4 +1,3 @@
-
 function BotonNavigate({handlerNavigate}) {
     
     return(
